Create Program1.c
This program takes a string from the user and converts all uppercase alphabetic characters into lowercase. It iterates through each character of the string and checks if it lies between ‘A’ and ‘Z’. If so, it adds 32 to convert it to the corresponding lowercase letter based on ASCII values. Finally, it displays the modified lowercase string as output.