Create Program1.c
This program copies the characters of a given string into another string in reverse order. It first moves a pointer to the end of the source string, stopping just before the null terminator. Then it traverses backward through the source string, copying each character into the destination string from left to right. After all characters are reversed and copied, a null terminator is added to mark the end of the new string. Finally, the reversed string is displayed.